backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
blinker==1.7.0
Brotli==1.0.9
certifi==2023.11.17
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
cycler==0.11.0
dataclasses==0.6
dataclasses-json==0.6.3
dbus-python==1.2.18
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.7.0
distro-info==1.0
dnspython==2.6.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
email_validator==2.2.0
environs==11.0.0
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi_cors==0.0.6
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.29.1
frozenlist==1.4.1
fs==2.4.12
fsspec==2023.12.2
gast==0.5.2
gotrue==0.2.0
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.4
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.20.2
idna==2.10
importlib-metadata==4.6.4
iniconfig==2.0.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama-index==0.9.34
lxml==4.8.0
lz4==3.1.3+dfsg
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib==3.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==8.10.0
mpmath==0.0.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.9
netifaces==0.11.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
openai==1.8.0
orjson==3.10.5
packaging==21.3
pandas==2.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pluggy==1.4.0
ply==3.11
postgrest-py==0.4.0
prompt-toolkit==3.0.40
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pypdf==4.0.0
pyrsistent==0.18.1
pytest==6.2.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.9
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
realtime-py==0.1.3
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
safetensors==0.4.1
scipy==1.8.0
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.37.2
supabase-py==0.0.2
sympy==1.9
systemd-python==234
tenacity==8.2.3
texttable==1.6.4
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
torch==2.1.2
tqdm==4.66.1
traitlets==5.13.0
transformers==4.36.2
triton==2.1.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-advantage-tools==8001
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.18
uvicorn==0.30.1
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.22.0
wcwidth==0.2.9
webencodings==0.5.1
websocket-client==1.2.3
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==1.0.0
